requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.7.0
boto3==1.29.4
botocore==1.32.4
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==41.0.4
cssselect2==0.7.0
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-storages==1.14.2
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.13.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
Markdown==3.6
MarkupSafe==2.1.5
msgpack==1.0.7
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
pipenv==2023.11.15
platformdirs==3.10.0
protobuf==4.25.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.2.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.4
virtualenv==20.25.0
webencodings==0.5.1
xhtml2pdf==0.2.14
zope.interface==6.1
